+++ scripts/Backstabs/player.lua
@@ -1,6 +1,10 @@
 local storage = require("openmw.storage")
 local ambient = require("openmw.ambient")
+local nearby = require("openmw.nearby")
+local input = require("openmw.input")
+local async = require("openmw.async")
 local core = require("openmw.core")
+local self = require("openmw.self")
 local ui = require("openmw.ui")
 local I = require('openmw.interfaces')
 require("scripts.Backstabs.backstabLogic")
@@ -9,6 +13,12 @@ local l10n = core.l10n("Backstabs")
 local sectionOnBackstab = storage.playerSection("SettingsBackstabs_onBackstab")
 local sectionToggles = storage.globalSection("SettingsBackstabs_toggles")
 
+input.registerActionHandler(input.actions.Sneak.key, async:callback(function()
+    for _, actor in pairs(nearby.actors) do
+        actor:sendEvent("playerSneaking", not self.controls.sneak)
+    end
+end))
+
 I.Combat.addOnHitHandler(function(attack)
     if not sectionToggles:get("playerCanBeBackstabbed") then return end
     DoBackstab(attack)
@@ -16,7 +26,9 @@ end)
 
 local function onLoad()
     -- always check your API version
-    if core.API_REVISION < 87 then ui.showMessage(l10n("messageOutdatedLuaAPI"), { showInDualogue = true }) end
+    if core.API_REVISION < 87 then
+        ui.showMessage(l10n("messageOutdatedLuaAPI"), { showInDualogue = true })
+    end
 end
 
 local function onBackstab(damageMult)
@@ -34,6 +46,6 @@ return {
         onLoad = onLoad
     },
     eventHandlers = {
-        onBackstab = onBackstab
+        BACKSTABS_onBackstab = onBackstab
     }
 }